handle creating dofferent reports content in a single module
